logging: remove Buffer template parameter type
This commit removes the Buffer template parameter type from the logger
and introduces a ByteBuffer interface, that the logger takes instead.

This increases the number of virtual calls, which is very unfortunate,
but it does allow the logger to be passed around without any need to
template the functions which take it. This was never going to be
accepted by the users, so it was needed.

In the future, I will have benchmarks for this and if it turns out that
those virtual calls are what is slowing us down, I'll instead embed the
buffer type directly into the logger and allow the user to decide which
one they want to use at compile time.